# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Telegram AI Autonomous System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /schedules/{id}:
    get:
      operationId: getSchedule
      tags: [schedules]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    patch:
      operationId: updateSchedule
      tags: [schedules]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── AI ENGINE ─────────────────────────────────────────
...
  /security/sanctions:
    get:
      operationId: listSanctions
      tags: [security]
      responses:
        "200":
          description: Sanctions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sanction"
    post:
      operationId: createSanction
      tags: [security]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SanctionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sanction"
  /security/sanctions/{id}:
    delete:
      operationId: deleteSanction
      tags: [security]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Schedule:
      type: object
      required: [id, name, cronExpression, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        campaignId:
          type: ["integer", "null"]
        cronExpression:
          type: string
        isActive:
          type: boolean
        lastRun:
          type: ["string", "null"]
        nextRun:
          type: ["string", "null"]
        runCount:
          type: integer
        createdAt:
          type: string
    ScheduleInput:
      type: object
      required: [name, cronExpression]
      properties:
        name:
          type: string
        campaignId:
          type: integer
        cronExpression:
          type: string
        isActive:
          type: boolean
    ScheduleUpdate:
      type: object
      properties:
        name:
          type: string
        cronExpression:
          type: string
        isActive:
          type: boolean
        campaignId:
          type: integer

    # ── AI ────────────────────────────────────────────────
...
    Sanction:
      type: object
      required: [id, type, targetUsername, reason, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [ban, mute, warn, kick]
        targetUsername:
          type: string
        targetUserId:
          type: ["string", "null"]
        groupId:
          type: ["integer", "null"]
        accountId:
          type: ["integer", "null"]
        reason:
          type: string
        expiresAt:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    SanctionInput:
      type: object
      required: [type, targetUsername, reason]
      properties:
        type:
          type: string
          enum: [ban, mute, warn, kick]
        targetUsername:
          type: string
        targetUserId:
          type: string
        groupId:
          type: integer
        accountId:
          type: integer
        reason:
          type: string
        expiresAt:
          type: string
...