# lib/api-zod/src/generated/types/scheduleUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram AI Autonomous System API
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleUpdate {
  name?: string;
  cronExpression?: string;
  isActive?: boolean;
  campaignId?: number;
}